initramfs

+++ initramfs.ownhere.2.3.4.CWM4/ownhere/scripts/init_4cleardatalog.sh
@@ -1,3 +1,5 @@
 #!/system/bin/sh
 
-rm -r /data/log/*
+if ! [ -e /data/log/disableautoclear ]; then
+	rm -r /data/log/*
+fi
